requirements.txt
apache-airflow==2.10.0
apache-airflow-providers-mongo==4.3.0
apache-airflow-providers-postgres==5.8.0
pymongo==4.6.1
pandas==2.1.4
psycopg2-binary==2.9.9
